+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-08-07 21:59:53
-// Commit: perf(i18n): apply RepaintBoundary on animated i18n widgets
+// Updated at: 2026-08-08 11:38:29
+// Commit: fix(i18n): fix memory leak by cancelling i18n subscription
 
 class i18nComponentState {
-  final String commitRef = '3448';
-  final String lastMessage = 'perf(i18n): apply RepaintBoundary on animated i18n widgets';
-  final DateTime timestamp = DateTime.parse('2026-08-07T21:59:53+05:00');
+  final String commitRef = '3454';
+  final String lastMessage = 'fix(i18n): fix memory leak by cancelling i18n subscription';
+  final DateTime timestamp = DateTime.parse('2026-08-08T11:38:29+05:00');
 }